Moving towards each submodule having its own test
Also, adding the setup.[cfg|py] files needed to
make this a proper package.
