+++ setup.cfg
@@ -1,0 +1,21 @@
+[metadata]
+name = eldonationtracker
+author = Eric Mesa
+author-email = eric@ericmesa.com
+url = http://djotaku.github.io/ELDonationTracker/
+version = v3.3
+description =  A Python-based donation tracker for Extra Life streams
+long_description = http://djotaku.github.io/ELDonationTracker/
+keywords = extra life, XSplit, obs, livestreaming, video games
+license = GPLv3
+classifiers = 
+    Intended Audience :: End Users/Desktop
+    License :: OSI Approved :: GNU General Public License v3 (GPLv3)
+    Programming Language :: Python :: 3
+
+[options]
+include_package_data = True
+packages = find:
+install_requires =
+    PyQt5
+    xdgenvpy
